requirements.txt
allure-pytest==2.11.1
allure-python-commons==2.11.1
pytest==7.2.0
requests==2.28.1
selenium==4.6.0